Complete UI Rework
Action_Box is a button with 4 states and a graphical sprite for each, which the classes Void_Box and Param_Box inherate from. These are buttons that given a function pointer will execute said function on click.

Param_Box goes a step further and gives a default parameter to said function. This parameter can be of any data type thanks to the class being a template and is stored in the button.

Ex: ShowWindow() would always be the same process with the window being opened depending on the parameter. Meaning that several buttons can be given this function to execute but a different window (the parameter) to open each.

Check_Box is a variation of a Void_Box, which has a bool* which his ON/OFF active state depends on. On click it will switch this state and execute a function if given one.

Ex: A button to activate/deactivate fullscreen or a button to turn on or off a game functionality.