Group improvements & Player fix
Group::SpreadDestinations() & Group::FindClosestWalkable() improved, they work now.
Now when a group is ordered to go to a non walkable tile, it goes to the closes walkable one.
Also, troops in the same group no longer stack in the same tile.

Also, since the constructor of the ModulePlayer caused some problems, i deleted it (it was the default one, so there's no problem with that).