Changed buildings and objects array name and its pool functionality
Previously, Buildings and Objects were working as pools, but they must'n because they are loaded all once before starting the game and remain the same during all game, so they don't need big changes in run time.

So they are not pools. I've changed it, they are just arrays and are loaded from the map and activated at entities