Little Fix on audio and unit
Just fixed the ControlVolume, which suposed to lower all volumes (SFX & Music) but didn't. Now it does.

I put the volume to 0 in scene just to not to hear always the same at each compilation and burnout xddd.

Also, there was an if in Unit that was empty and make the game crash, so I deleted it.