Hordes Fixed
I had to do make a function to clean the hordes from communist units which is called each 2 seconds as well as the check to change the round

Horde Manager module had no name, now it has xd.

I have also did the module Horde Manager to have a correct code of Public and Private variables and functions and changed the function hordeRoundto() to restartRounds(), because it was pretty a mess to change the round number, so it just restart the rounds to 0.